wip: nextauth middleware
