+++ app/client-v2/apps/goat/app/auth/layout.tsx
@@ -3,15 +3,15 @@
 import { makeStyles } from "@/lib/theme";
 
 import Box from "@p4b/ui/components/Box";
+import { Card } from "@p4b/ui/components/Card";
 import Grid from "@p4b/ui/components/Grid";
 
-interface OnboardingLayoutProps {
+interface AuthLayoutProps {
   children: React.ReactNode;
 }
 
-const OnboardingLayout = (props: OnboardingLayoutProps) => {
+const AuthLayout = (props: AuthLayoutProps) => {
   const { children } = props;
-
   const { classes } = useStyles();
 
   return (
@@ -42,7 +42,14 @@ const OnboardingLayout = (props: OnboardingLayoutProps) => {
               />
             </Box>
           </Box>
-          <div className={classes.child}>{children}</div>
+
+          <div className={classes.child}>
+            <div className={classes.cardContainer}>
+              <Card width={480} noHover={true} className={classes.paper}>
+                {children}
+              </Card>
+            </div>
+          </div>
         </Grid>
         <Grid xs={12} lg={6} className={classes.gridRight}>
           <Box sx={{ p: 3, width: 350 }} component="div">
@@ -58,12 +65,16 @@ const OnboardingLayout = (props: OnboardingLayoutProps) => {
   );
 };
 
-const useStyles = makeStyles({ name: { OnboardingLayout } })((theme) => ({
+const useStyles = makeStyles({ name: { AuthLayout } })((theme) => ({
   root: {
     display: "flex",
   },
   gridContainer: {
     flex: "1 1 auto",
+  },
+  cardContainer: {
+    display: "flex",
+    justifyContent: "center",
   },
   child: {
     width: "100%",
@@ -77,6 +88,13 @@ const useStyles = makeStyles({ name: { OnboardingLayout } })((theme) => ({
     width: "50%",
     backgroundImage: 'url("https://assets.plan4better.de/img/login/artwork_1.png")',
     height: "100vh",
+  },
+  paper: {
+    padding: theme.spacing(5),
+    width: 490,
+    height: "fit-content",
+    marginBottom: theme.spacing(4),
+    borderRadius: 4,
   },
   gridLeft: {
     height: "100vh",
@@ -98,4 +116,4 @@ const useStyles = makeStyles({ name: { OnboardingLayout } })((theme) => ({
   },
 }));
 
-export default OnboardingLayout;
+export default AuthLayout;
