+++ app/client-v2/apps/goat/app/auth/logout/page.tsx
@@ -1,0 +1,7 @@
+export default function Logout() {
+  return (
+    <div>
+      <h1>Logout</h1>
+    </div>
+  );
+}
