--- app/client-v2/apps/goat/app/onboarding/page.tsx
@@ -1,5 +1,0 @@
-import { redirect } from "next/navigation";
-
-export default async function OnBoardingPage({}) {
-  return redirect("/onboarding/organization");
-}
