+++ app/client-v2/apps/goat/middleware.ts
@@ -1,0 +1,25 @@
+import type { NextRequestWithAuth } from "next-auth/middleware";
+import { withAuth } from "next-auth/middleware";
+import { NextResponse } from "next/server";
+
+export default withAuth(
+  function middleware(request: NextRequestWithAuth) {
+    if (!request.nextauth.token) {
+      const url = new URL(`/api/auth/signin`, request.url);
+      url.searchParams.set("callbackUrl", encodeURI(request.url));
+      return NextResponse.redirect(url);
+    }
+
+    if (!request.nextauth.token?.user?.org_id) {
+      const url = new URL("/auth/organization", request.url);
+      return NextResponse.redirect(url);
+    }
+  },
+  {
+    callbacks: {
+      authorized: ({ token }) => !!token,
+    },
+  }
+);
+
+export const config = { matcher: ["/home", "/content", "/help", "/settings"] };
