+++ app/client-v2/packages/ui/components/Link.tsx
@@ -1,0 +1,3 @@
+import Link from "@mui/material/Link";
+
+export default Link;
